:recycle: 클래스 네이밍 변경
- study -> gathering, free -> playground
- post -> create
