+++ noriter/src/main/java/com/codewarts/noriter/article/dto/free/PlaygroundDetailResponse.java
@@ -2,6 +2,8 @@ package com.codewarts.noriter.article.dto.free;
 
 import com.codewarts.noriter.article.domain.Article;
 import com.codewarts.noriter.article.domain.Hashtag;
+import com.codewarts.noriter.comment.dto.comment.CommentResponse;
+import com.codewarts.noriter.member.dto.WriterInfoResponse;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,14 +13,12 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class FreeListResponse {
+public class PlaygroundDetailResponse {
 
     private Long id;
     private String title;
     private String content;
-
-    private String writerNickname;
-
+    private WriterInfoResponse writer;
     private boolean sameWriter;
     private List<String> hashtags;
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
@@ -27,23 +27,23 @@ public class FreeListResponse {
     private LocalDateTime editedTime;
     private boolean wish;
     private int wishCount;
-    private int commentCount;
-    private boolean completed;
+    private List<CommentResponse> comment;
 
-    public FreeListResponse(Article article, boolean sameWriter, boolean wish) {
+    public PlaygroundDetailResponse(Article article, boolean sameWriter, boolean wish) {
         this.id = article.getId();
         this.title = article.getTitle();
         this.content = article.getContent();
-        this.writerNickname = article.getWriter().getNickname();
+        this.writer = new WriterInfoResponse(article.getWriter());
         this.sameWriter = sameWriter;
         this.hashtags = article.getHashtags().stream()
             .map(Hashtag::getContent)
             .collect(Collectors.toList());
         this.writtenTime = article.getWrittenTime();
-        this.editedTime = article.getWrittenTime();
+        this.editedTime = article.getEditedTime();
         this.wish = wish;
         this.wishCount = article.getWishList().size();
-        this.commentCount = article.getComments().size();
-        this.completed = isCompleted();
+        this.comment = article.getComments().stream()
+            .map(CommentResponse::new)
+            .collect(Collectors.toList());
     }
 }
