+++ noriter/src/main/java/com/codewarts/noriter/article/dto/study/GatheringCreateRequest.java
@@ -1,36 +1,33 @@
-package com.codewarts.noriter.article.dto.question;
+package com.codewarts.noriter.article.dto.study;
 
-import com.codewarts.noriter.article.domain.Question;
+import com.codewarts.noriter.article.domain.Study;
 import com.codewarts.noriter.article.domain.type.ArticleType;
 import com.codewarts.noriter.article.domain.type.StatusType;
 import com.codewarts.noriter.member.domain.Member;
 import java.time.LocalDateTime;
 import java.util.List;
 import javax.validation.constraints.NotBlank;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-@AllArgsConstructor
-public class QuestionPostRequest {
+public class GatheringCreateRequest {
 
     @NotBlank(message = "제목은 필수입니다.")
     private String title;
     @NotBlank(message = "내용은 필수입니다.")
     private String content;
-
     private List<String> hashtags;
 
-    public Question toEntity(Member writer) {
-        return Question.builder()
+    public Study toEntity(Member writer) {
+        return Study.builder()
             .title(title)
             .content(content)
             .writer(writer)
             .writtenTime(LocalDateTime.now())
             .editedTime(LocalDateTime.now())
-            .articleType(ArticleType.QUESTION)
+            .articleType(ArticleType.STUDY)
             .status(StatusType.INCOMPLETE)
             .build();
     }
