+++ noriter/src/main/java/com/codewarts/noriter/article/dto/study/GatheringUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.codewarts.noriter.article.dto.free;
+package com.codewarts.noriter.article.dto.study;
 
 import java.util.List;
 import javax.validation.constraints.NotBlank;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class FreeEditRequest {
+public class GatheringUpdateRequest {
 
     @NotBlank(message = "제목은 필수입니다.")
     private String title;
