+++ noriter/src/main/java/com/codewarts/noriter/article/repository/GatheringRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface StudyRepository extends JpaRepository<Study, Long> {
+public interface GatheringRepository extends JpaRepository<Study, Long> {
 
     @Query("select s from Study s where s.status = :status")
     List<Study> findStudyByCompleted(@Param("status") StatusType status);
