+++ noriter/src/main/java/com/codewarts/noriter/comment/dto/comment/CommentCreateRequest.java
@@ -1,33 +1,31 @@
-package com.codewarts.noriter.article.dto.free;
+package com.codewarts.noriter.comment.dto.comment;
 
 import com.codewarts.noriter.article.domain.Article;
-import com.codewarts.noriter.article.domain.type.ArticleType;
+import com.codewarts.noriter.comment.domain.Comment;
 import com.codewarts.noriter.member.domain.Member;
 import java.time.LocalDateTime;
-import java.util.List;
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class FreePostRequest {
+public class CommentCreateRequest {
 
-    @NotBlank(message = "제목은 필수입니다.")
-    private String title;
     @NotBlank(message = "내용은 필수입니다.")
     private String content;
-    private List<String> hashtags;
+    @NotNull(message = "내용은 필수입니다.")
+    private Boolean secret;
 
-    public Article toEntity(Member writer) {
-        return Article.builder()
-            .title(title)
-            .content(content)
+    public Comment toEntity(Article article, Member writer) {
+        return Comment.builder()
+            .article(article)
             .writer(writer)
+            .content(content)
+            .secret(secret)
             .writtenTime(LocalDateTime.now())
             .editedTime(LocalDateTime.now())
-            .articleType(ArticleType.FREE)
             .build();
     }
 }
-
