+++ noriter/src/test/java/com/codewarts/noriter/article/docs/gathering/GatheringDeleteTest.java
@@ -1,7 +1,6 @@
-package com.codewarts.noriter.article.docs.study;
+package com.codewarts.noriter.article.docs.gathering;
 
 import static com.codewarts.noriter.exception.type.ArticleExceptionType.ARTICLE_NOT_MATCHED_WRITER;
-import static com.codewarts.noriter.exception.type.CommonExceptionType.INVALID_REQUEST;
 import static com.codewarts.noriter.exception.type.MemberExceptionType.MEMBER_NOT_FOUND;
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
@@ -22,8 +21,6 @@ import com.codewarts.noriter.exception.type.CommonExceptionType;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
-import java.util.List;
-import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
@@ -42,7 +39,7 @@ import org.springframework.test.context.jdbc.Sql;
 @ExtendWith({RestDocumentationExtension.class})
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql("classpath:/data.sql")
-class StudyEditTest {
+class GatheringDeleteTest {
 
     @LocalServerPort
     int port;
@@ -72,84 +69,67 @@ class StudyEditTest {
     }
 
     @Test
-    void 글을_수정한다() {
+    void 게시글을_삭제한다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
-
-        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
-            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-            .body(requestBody)
+            .pathParam("id", 2)
 
             .when()
-            .put("/community/gathering/{id}")
+            .delete("/community/gathering/{id}")
 
             .then()
             .statusCode(HttpStatus.OK.value());
     }
-
     @Test
     void Path_Variable이_없는_경우_예외를_발생시킨다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
-        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
-            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
-
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", " ")
-            .body(requestBody)
+            .pathParam("id", "  ")
 
             .when()
-            .put("/community/gathering/{id}")
+            .delete("/community/gathering/{id}")
 
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("gatheringEdit.id: ID가 비어있습니다."));
+            .body("message", equalTo("remove.id: ID가 비어있습니다."));
     }
 
     @Test
     void id가_유효하지_않는_경우_예외를_발생시킨다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
-        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
-            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
-
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
             .pathParam("id", -1)
-            .body(requestBody)
 
             .when()
-            .put("/community/gathering/{id}")
+            .delete("/community/gathering/{id}")
 
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("gatheringEdit.id: 게시글 ID는 양수이어야 합니다."));
+            .body("message", equalTo("remove.id: 게시글 ID는 양수이어야 합니다."));
     }
 
     @Test
     void id가_존재하지_않는_경우_예외를_발생시킨다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
-        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
-            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
-
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 999999999)
-            .body(requestBody)
+            .pathParam("id", 999999)
 
             .when()
-            .put("/community/gathering/{id}")
+            .delete("/community/gathering/{id}")
 
             .then()
             .statusCode(ArticleExceptionType.ARTICLE_NOT_FOUND.getStatus().value())
@@ -158,44 +138,16 @@ class StudyEditTest {
     }
 
     @Test
-    void 필수값이_비어있을_경우_예외가_발생한다() {
-        String accessToken = jwtProvider.issueAccessToken(1L);
-
-        Map<String, Object> requestBody = Map.of("content", "하나둘셋 얍",
-            "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
+    void 존재하지_않는_회원인_경우_예외가_발생한다() {
+        String accessToken = jwtProvider.issueAccessToken(99999L);
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
             .pathParam("id", 1)
-            .body(requestBody)
 
             .when()
-            .put("/community/gathering/{id}")
-
-            .then()
-            .statusCode(INVALID_REQUEST.getStatus().value())
-            .body("errorCode", equalTo(INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo(INVALID_REQUEST.getErrorMessage()))
-            .body("detail.title", equalTo("제목은 필수입니다."));
-    }
-
-    @Test
-    void 존재하지_않는_회원인_경우_예외가_발생한다() {
-        String accessToken = jwtProvider.issueAccessToken(99999999L);
-
-        Map<String, Object> requestBody = Map.of("title", "안녕하세용",
-            "content", "헬륨가스를모곳지", "hashtags",
-            List.of("자유게시판", "개발자좋아효", "코린이"));
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 10)
-            .body(requestBody)
-
-            .when()
-            .put("/community/gathering/{id}")
+            .delete("/community/gathering/{id}")
 
             .then()
             .statusCode(MEMBER_NOT_FOUND.getStatus().value())
@@ -207,18 +159,13 @@ class StudyEditTest {
     void 작성자가_일치하지_않는_경우_예외가_발생한다() {
         String accessToken = jwtProvider.issueAccessToken(2L);
 
-        Map<String, Object> requestBody = Map.of("title", "안녕하세용",
-            "content", "헬륨가스를모곳지", "hashtags",
-            List.of("자유게시판", "개발자좋아효", "코린이"));
-
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 2)
-            .body(requestBody)
+            .pathParam("id", 1)
 
             .when()
-            .put("/community/gathering/{id}")
+            .delete("/community/gathering/{id}")
 
             .then()
             .statusCode(ARTICLE_NOT_MATCHED_WRITER.getStatus().value())
