+++ noriter/src/test/java/com/codewarts/noriter/article/docs/gathering/GatheringDetailTest.java
@@ -1,4 +1,4 @@
-package com.codewarts.noriter.article.docs.study;
+package com.codewarts.noriter.article.docs.gathering;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
@@ -39,7 +39,7 @@ import org.springframework.test.context.jdbc.Sql;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Profile({"test"})
 @Sql("classpath:/data.sql")
-public class StudyDetailTest {
+public class GatheringDetailTest {
 
     @LocalServerPort
     int port;
@@ -151,7 +151,7 @@ public class StudyDetailTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("gatheringDetail.id: ID가 비어있습니다."));
+            .body("message", equalTo("getDetail.id: ID가 비어있습니다."));
     }
 
     @Test
@@ -167,7 +167,7 @@ public class StudyDetailTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("gatheringDetail.id: 게시글 ID는 양수이어야 합니다."));
+            .body("message", equalTo("getDetail.id: 게시글 ID는 양수이어야 합니다."));
     }
 
     @Test
