+++ noriter/src/test/java/com/codewarts/noriter/article/docs/gathering/GatheringUpdateTest.java
@@ -1,8 +1,6 @@
-package com.codewarts.noriter.article.docs.question;
+package com.codewarts.noriter.article.docs.gathering;
 
 import static com.codewarts.noriter.exception.type.ArticleExceptionType.ARTICLE_NOT_MATCHED_WRITER;
-import static com.codewarts.noriter.exception.type.AuthExceptionType.EMPTY_ACCESS_TOKEN;
-import static com.codewarts.noriter.exception.type.AuthExceptionType.TAMPERED_ACCESS_TOKEN;
 import static com.codewarts.noriter.exception.type.CommonExceptionType.INVALID_REQUEST;
 import static com.codewarts.noriter.exception.type.MemberExceptionType.MEMBER_NOT_FOUND;
 import static io.restassured.RestAssured.given;
@@ -18,12 +16,7 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.removeHeaders;
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
 
-import com.codewarts.noriter.article.domain.Hashtag;
-import com.codewarts.noriter.article.domain.Question;
-import com.codewarts.noriter.article.dto.question.QuestionUpdateRequest;
-import com.codewarts.noriter.article.repository.QuestionRepository;
 import com.codewarts.noriter.auth.jwt.JwtProvider;
-import com.codewarts.noriter.exception.GlobalNoriterException;
 import com.codewarts.noriter.exception.type.ArticleExceptionType;
 import com.codewarts.noriter.exception.type.CommonExceptionType;
 import io.restassured.RestAssured;
@@ -31,7 +24,6 @@ import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
@@ -44,19 +36,17 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.restdocs.RestDocumentationContextProvider;
 import org.springframework.restdocs.RestDocumentationExtension;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.test.context.jdbc.Sql;
 
-
-@Transactional
 @DisplayNameGeneration(ReplaceUnderscores.class)
 @ExtendWith({RestDocumentationExtension.class})
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class QuestionEditTest {
+@Sql("classpath:/data.sql")
+class GatheringUpdateTest {
 
     @LocalServerPort
     int port;
-    @Autowired
-    private QuestionRepository questionRepository;
+
     @Autowired
     protected JwtProvider jwtProvider;
 
@@ -82,125 +72,11 @@ class QuestionEditTest {
     }
 
     @Test
-    void 제목을_수정한다() {
+    void 글을_수정한다() {
+        String accessToken = jwtProvider.issueAccessToken(1L);
 
-        Question question = questionRepository.findById(6L)
-            .orElseThrow(() -> new GlobalNoriterException(ArticleExceptionType.ARTICLE_NOT_FOUND));
-        String accessToken = jwtProvider.issueAccessToken(question.getWriter().getId());
-        List<String> hashtag = question.getHashtags().stream().map(Hashtag::getContent)
-            .collect(Collectors.toList());
-        QuestionUpdateRequest updateRequest = new QuestionUpdateRequest("수정된 제목", question.getContent(), hashtag);
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
-            .body(updateRequest)
-
-            .when()
-            .put("/community/question/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-    @Test
-    void 내용을_수정한다() {
-
-        Question question = questionRepository.findById(6L)
-            .orElseThrow(() -> new GlobalNoriterException(ArticleExceptionType.ARTICLE_NOT_FOUND));
-        String accessToken = jwtProvider.issueAccessToken(question.getWriter().getId());
-        List<String> hashtag = question.getHashtags().stream().map(Hashtag::getContent)
-            .collect(Collectors.toList());
-
-        QuestionUpdateRequest updateRequest = new QuestionUpdateRequest(question.getTitle(), "수정된 내용", hashtag);
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
-            .body(updateRequest)
-
-            .when()
-            .put("/community/question/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void 해시태그를_수정한다() {
-
-        Question question = questionRepository.findById(6L)
-            .orElseThrow(RuntimeException::new);
-        String accessToken = jwtProvider.issueAccessToken(question.getWriter().getId());
-
-        QuestionUpdateRequest updateRequest = new QuestionUpdateRequest(question.getTitle(), question.getContent(), List.of("수정 태그1", "수정 태그2"));
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
-            .body(updateRequest)
-
-            .when()
-            .put("/community/question/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void 전부_수정한다() {
-
-        Question question = questionRepository.findById(6L)
-            .orElseThrow(RuntimeException::new);
-        String accessToken = jwtProvider.issueAccessToken(question.getWriter().getId());
-
-        QuestionUpdateRequest updateRequest = new QuestionUpdateRequest("수정된 제목", "수정된 내용", List.of("수정 태그1", "수정 태그2"));
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
-            .body(updateRequest)
-
-            .when()
-            .put("/community/question/{id}")
-
-            .then()
-            .statusCode(HttpStatus.OK.value());
-    }
-
-    @Test
-    void Access_Token이_비어있는_경우_예외_발생() {
-        String accessToken = " ";
-
-        Map<String, Object> requestBody = Map.of("title", "질문있어요",
-            "content", "스프링 어려워요", "hashtags",
-            List.of("질문게시판", "고수가 되고싶어요", "코린이"));
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
-            .body(requestBody)
-
-            .when()
-            .put("/community/question/{id}")
-
-            .then()
-            .statusCode(EMPTY_ACCESS_TOKEN.getStatus().value())
-            .body("errorCode", equalTo(EMPTY_ACCESS_TOKEN.getErrorCode()))
-            .body("message", equalTo(EMPTY_ACCESS_TOKEN.getErrorMessage()));
-    }
-
-    @Test
-    void Access_Token이_변조된_경우_예외_발생() {
-        String accessToken = jwtProvider.issueAccessToken(1L) + "123";
-
-        Map<String, Object> requestBody = Map.of("title", "질문있어요",
-            "content", "스프링 어려워요", "hashtags",
-            List.of("질문게시판", "고수가 되고싶어요", "코린이"));
+        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
+            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
@@ -209,12 +85,10 @@ class QuestionEditTest {
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
-            .statusCode(TAMPERED_ACCESS_TOKEN.getStatus().value())
-            .body("errorCode", equalTo(TAMPERED_ACCESS_TOKEN.getErrorCode()))
-            .body("message", equalTo(TAMPERED_ACCESS_TOKEN.getErrorMessage()));
+            .statusCode(HttpStatus.OK.value());
     }
 
     @Test
@@ -231,12 +105,12 @@ class QuestionEditTest {
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("questionEdit.id: ID가 비어있습니다."));
+            .body("message", equalTo("edit.id: ID가 비어있습니다."));
     }
 
     @Test
@@ -253,12 +127,12 @@ class QuestionEditTest {
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("questionEdit.id: 게시글 ID는 양수이어야 합니다."));
+            .body("message", equalTo("edit.id: 게시글 ID는 양수이어야 합니다."));
     }
 
     @Test
@@ -275,7 +149,7 @@ class QuestionEditTest {
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
             .statusCode(ArticleExceptionType.ARTICLE_NOT_FOUND.getStatus().value())
@@ -293,11 +167,11 @@ class QuestionEditTest {
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
+            .pathParam("id", 1)
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
             .statusCode(INVALID_REQUEST.getStatus().value())
@@ -317,11 +191,11 @@ class QuestionEditTest {
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
+            .pathParam("id", 10)
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
             .statusCode(MEMBER_NOT_FOUND.getStatus().value())
@@ -340,14 +214,15 @@ class QuestionEditTest {
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 6)
+            .pathParam("id", 2)
             .body(requestBody)
 
             .when()
-            .put("/community/question/{id}")
+            .put("/community/gathering/{id}")
 
             .then()
             .statusCode(ARTICLE_NOT_MATCHED_WRITER.getStatus().value())
             .body("errorCode", equalTo(ARTICLE_NOT_MATCHED_WRITER.getErrorCode()));
     }
+
 }
