+++ noriter/src/test/java/com/codewarts/noriter/article/docs/playground/PlaygroundCreateTest.java
@@ -1,4 +1,5 @@
-package com.codewarts.noriter.article.docs.study;
+package com.codewarts.noriter.article.docs.playground;
+
 
 import static com.codewarts.noriter.exception.type.CommonExceptionType.INVALID_REQUEST;
 import static com.codewarts.noriter.exception.type.MemberExceptionType.MEMBER_NOT_FOUND;
@@ -41,7 +42,8 @@ import org.springframework.test.context.jdbc.Sql;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Profile({"test"})
 @Sql("classpath:/data.sql")
-class StudyCreateTest {
+public class PlaygroundCreateTest {
+
 
     @LocalServerPort
     int port;
@@ -75,7 +77,8 @@ class StudyCreateTest {
         String accessToken = jwtProvider.issueAccessToken(2L);
 
         Map<String, Object> requestBody = Map.of("title", "안녕하세용",
-            "content", "헬륨가스를모곳지", "hashtags", List.of("SPRING", "JPA", "코린이"));
+            "content", "헬륨가스를모곳지", "hashtags",
+            List.of("자유게시판", "개발자좋아효", "코린이"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
@@ -83,7 +86,7 @@ class StudyCreateTest {
             .body(requestBody)
 
             .when()
-            .post("/community/gathering")
+            .post("/community/playground")
 
             .then()
             .statusCode(HttpStatus.OK.value());
@@ -93,8 +96,8 @@ class StudyCreateTest {
     void 필수값이_비어있을_경우_예외가_발생한다() {
         String accessToken = jwtProvider.issueAccessToken(2L);
 
-        Map<String, Object> requestBody = Map.of("title", "안녕하세용",
-            "hashtags", List.of("SPRING", "JPA", "코린이"));
+        Map<String, Object> requestBody = Map.of("content", "헬륨가스를모곳지", "hashtags",
+            List.of("자유게시판", "개발자좋아효", "코린이"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
@@ -102,20 +105,22 @@ class StudyCreateTest {
             .body(requestBody)
 
             .when()
-            .post("/community/gathering")
+            .post("/community/playground")
 
             .then()
+            .statusCode(INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(INVALID_REQUEST.getErrorCode()))
             .body("message", equalTo(INVALID_REQUEST.getErrorMessage()))
-            .body("detail.content", equalTo("내용은 필수입니다."));
+            .body("detail.title", equalTo("제목은 필수입니다."));
     }
 
     @Test
     void 존재하지_않는_회원인_경우_예외가_발생한다() {
-        String accessToken = jwtProvider.issueAccessToken(222222L);
+        String accessToken = jwtProvider.issueAccessToken(99999999L);
 
         Map<String, Object> requestBody = Map.of("title", "안녕하세용",
-            "content", "헬륨가스를모곳지", "hashtags", List.of("SPRING", "JPA", "코린이"));
+            "content", "헬륨가스를모곳지", "hashtags",
+            List.of("자유게시판", "개발자좋아효", "코린이"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
@@ -123,12 +128,11 @@ class StudyCreateTest {
             .body(requestBody)
 
             .when()
-            .post("/community/gathering")
+            .post("/community/playground")
 
             .then()
             .statusCode(MEMBER_NOT_FOUND.getStatus().value())
             .body("errorCode", equalTo(MEMBER_NOT_FOUND.getErrorCode()))
             .body("message", equalTo(MEMBER_NOT_FOUND.getErrorMessage()));
     }
-
 }
