+++ noriter/src/test/java/com/codewarts/noriter/article/docs/playground/PlaygroundDetailTest.java
@@ -1,4 +1,4 @@
-package com.codewarts.noriter.article.docs.free;
+package com.codewarts.noriter.article.docs.playground;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
@@ -40,7 +40,7 @@ import org.springframework.test.context.jdbc.Sql;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Profile({"test"})
 @Sql("classpath:/data.sql")
-class FreeDetailTest {
+class PlaygroundDetailTest {
 
     @LocalServerPort
     int port;
@@ -150,7 +150,7 @@ class FreeDetailTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("freeDetail.id: 게시글 ID는 양수이어야 합니다."));
+            .body("message", equalTo("getDetail.id: 게시글 ID는 양수이어야 합니다."));
     }
 
     @Test
@@ -166,7 +166,7 @@ class FreeDetailTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("freeDetail.id: ID가 비어있습니다."));
+            .body("message", equalTo("getDetail.id: ID가 비어있습니다."));
     }
 
     @Test
