+++ noriter/src/test/java/com/codewarts/noriter/article/docs/playground/PlaygroundListTest.java
@@ -1,4 +1,4 @@
-package com.codewarts.noriter.article.docs.study;
+package com.codewarts.noriter.article.docs.playground;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.Matchers.equalTo;
@@ -14,7 +14,6 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.re
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
 
 import com.codewarts.noriter.auth.jwt.JwtProvider;
-import com.codewarts.noriter.exception.type.CommonExceptionType;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
@@ -38,7 +37,7 @@ import org.springframework.test.context.jdbc.Sql;
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Profile({"test"})
 @Sql("classpath:/data.sql")
-class StudyListTest {
+class PlaygroundListTest {
 
     @LocalServerPort
     int port;
@@ -72,26 +71,26 @@ class StudyListTest {
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
-            .param("status", "INCOMPLETE")
 
             .when()
-            .get("/community/gathering")
+            .get("/community/playground")
 
             .then()
             .statusCode(HttpStatus.OK.value())
-            .body("[0].id", equalTo(1))
-            .body("[0].title", equalTo("테스트를 해볼것이당"))
-            .body("[0].content", equalTo("안녕하냐고오옹"))
-            .body("[0].writerNickname", equalTo("admin1"))
+            .body("[0].id", equalTo(10))
+            .body("[0].title", equalTo("붕어빵 먹고싶어요"))
+            .body("[0].content", equalTo("강남 붕어빵 맛잇는 집"))
+            .body("[0].writerNickname", equalTo("admin2"))
             .body("[0].sameWriter", equalTo(false))
-            .body("[0].writtenTime", equalTo("2022-11-11 16:25:58"))
-            .body("[0].editedTime", equalTo("2022-11-11 16:25:58"))
-            .body("[0].hashtags[0]", equalTo("SPRING"))
-            .body("[0].hashtags[1]", equalTo("JPA"))
+            .body("[0].writtenTime", equalTo("2022-11-25 16:25:58"))
+            .body("[0].editedTime", equalTo("2022-11-25 16:25:58"))
+            .body("[0].hashtags[0]", equalTo("강남역"))
+            .body("[0].hashtags[1]", equalTo("붕어팥"))
             .body("[0].wish", equalTo(false))
             .body("[0].wishCount", equalTo(1))
-            .body("[0].commentCount", equalTo(4));
+            .body("[0].commentCount", equalTo(1));
     }
+
     @Test
     void 로그인_후_리스트를_조회한다() {
         String accessToken = jwtProvider.issueAccessToken(2L);
@@ -99,56 +98,23 @@ class StudyListTest {
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .param("status", "INCOMPLETE")
 
             .when()
-            .get("/community/gathering")
+            .get("/community/playground")
 
             .then()
             .statusCode(HttpStatus.OK.value())
-            .body("[0].id", equalTo(1))
-            .body("[0].title", equalTo("테스트를 해볼것이당"))
-            .body("[0].content", equalTo("안녕하냐고오옹"))
-            .body("[0].writerNickname", equalTo("admin1"))
-            .body("[0].sameWriter", equalTo(false))
-            .body("[0].writtenTime", equalTo("2022-11-11 16:25:58"))
-            .body("[0].editedTime", equalTo("2022-11-11 16:25:58"))
-            .body("[0].hashtags[0]", equalTo("SPRING"))
-            .body("[0].hashtags[1]", equalTo("JPA"))
+            .body("[0].id", equalTo(10))
+            .body("[0].title", equalTo("붕어빵 먹고싶어요"))
+            .body("[0].content", equalTo("강남 붕어빵 맛잇는 집"))
+            .body("[0].writerNickname", equalTo("admin2"))
+            .body("[0].sameWriter", equalTo(true))
+            .body("[0].writtenTime", equalTo("2022-11-25 16:25:58"))
+            .body("[0].editedTime", equalTo("2022-11-25 16:25:58"))
+            .body("[0].hashtags[0]", equalTo("강남역"))
+            .body("[0].hashtags[1]", equalTo("붕어팥"))
             .body("[0].wish", equalTo(true))
             .body("[0].wishCount", equalTo(1))
-            .body("[0].commentCount", equalTo(4));
+            .body("[0].commentCount", equalTo(1));
     }
-
-    @Test
-    void 잘못된_requestParam_값인_경우_예외를_발생시킨다() {
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .param("comp", "INCOMPLETE")
-
-            .when()
-            .get("/community/gathering")
-
-            .then()
-            .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
-            .body("errorCode", equalTo(CommonExceptionType.INCORRECT_REQUEST_PARAM.getErrorCode()))
-            .body("message", equalTo(CommonExceptionType.INCORRECT_REQUEST_PARAM.getErrorMessage()));
-    }
-    @Test
-    void 잘못된_requestParam_타입인_경우_예외를_발생시킨다() {
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .param("status", "dd")
-
-            .when()
-            .get("/community/gathering")
-
-            .then()
-            .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
-            .body("errorCode", equalTo(CommonExceptionType.INCORRECT_REQUEST_VALUE.getErrorCode()))
-            .body("message", equalTo(CommonExceptionType.INCORRECT_REQUEST_VALUE.getErrorMessage()));
-    }
-
 }
