+++ noriter/src/test/java/com/codewarts/noriter/article/docs/playground/PlaygroundUpdateTest.java
@@ -1,9 +1,6 @@
-package com.codewarts.noriter.article.docs.study;
+package com.codewarts.noriter.article.docs.playground;
 
-import static com.codewarts.noriter.exception.type.ArticleExceptionType.ALREADY_CHANGED_STATUS;
-import static com.codewarts.noriter.exception.type.ArticleExceptionType.ARTICLE_NOT_FOUND;
 import static com.codewarts.noriter.exception.type.ArticleExceptionType.ARTICLE_NOT_MATCHED_WRITER;
-import static com.codewarts.noriter.exception.type.CommonExceptionType.INCORRECT_REQUEST_VALUE;
 import static com.codewarts.noriter.exception.type.CommonExceptionType.INVALID_REQUEST;
 import static com.codewarts.noriter.exception.type.MemberExceptionType.MEMBER_NOT_FOUND;
 import static io.restassured.RestAssured.given;
@@ -20,10 +17,13 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.re
 import static org.springframework.restdocs.restassured3.RestAssuredRestDocumentation.documentationConfiguration;
 
 import com.codewarts.noriter.auth.jwt.JwtProvider;
+import com.codewarts.noriter.exception.type.ArticleExceptionType;
+import com.codewarts.noriter.exception.type.CommonExceptionType;
 import io.restassured.RestAssured;
 import io.restassured.builder.RequestSpecBuilder;
 import io.restassured.specification.RequestSpecification;
-import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
@@ -42,7 +42,7 @@ import org.springframework.test.context.jdbc.Sql;
 @ExtendWith({RestDocumentationExtension.class})
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql("classpath:/data.sql")
-class StudyUpdateStatusTest {
+class PlaygroundUpdateTest {
 
     @LocalServerPort
     int port;
@@ -72,17 +72,20 @@ class StudyUpdateStatusTest {
     }
 
     @Test
-    void 게시글의_모집상태를_변경한다() {
+    void 글을_수정한다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
+
+        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
+            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-            .body(Collections.singletonMap("status", "complete"))
+            .pathParam("id", 2)
+            .body(requestBody)
 
             .when()
-            .patch("/community/gathering/{id}")
+            .put("/community/playground/{id}")
 
             .then()
             .statusCode(HttpStatus.OK.value());
@@ -92,72 +95,107 @@ class StudyUpdateStatusTest {
     void Path_Variable이_없는_경우_예외를_발생시킨다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
+        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
+            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
+
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
             .pathParam("id", " ")
-            .body(Collections.singletonMap("status", "complete"))
+            .body(requestBody)
 
             .when()
-            .patch("/community/gathering/{id}")
+            .put("/community/playground/{id}")
 
             .then()
-            .statusCode(INVALID_REQUEST.getStatus().value())
-            .body("errorCode", equalTo(INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("recruitmentCompletionUpdate.id: ID가 비어있습니다."));
+            .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
+            .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
+            .body("message", equalTo("edit.id: ID가 비어있습니다."));
     }
 
     @Test
     void id가_유효하지_않는_경우_예외를_발생시킨다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
+        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
+            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
+
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
             .pathParam("id", -1)
-            .body(Collections.singletonMap("status", "complete"))
+            .body(requestBody)
 
             .when()
-            .patch("/community/gathering/{id}")
+            .put("/community/playground/{id}")
 
             .then()
-            .statusCode(INVALID_REQUEST.getStatus().value())
-            .body("errorCode", equalTo(INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("recruitmentCompletionUpdate.id: 게시글 ID는 양수이어야 합니다."));
+            .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
+            .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
+            .body("message", equalTo("edit.id: 게시글 ID는 양수이어야 합니다."));
+    }
+
+    @Test
+    void id가_존재하지_않는_경우_예외를_발생시킨다() {
+        String accessToken = jwtProvider.issueAccessToken(1L);
+
+        Map<String, Object> requestBody = Map.of("title", "내가 글을 수정해볼게",
+            "content", "하나둘셋 얍", "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
+
+        given(documentationSpec)
+            .contentType(APPLICATION_JSON_VALUE)
+            .header(AUTHORIZATION, accessToken)
+            .pathParam("id", 999999999)
+            .body(requestBody)
+
+            .when()
+            .put("/community/playground/{id}")
+
+            .then()
+            .statusCode(ArticleExceptionType.ARTICLE_NOT_FOUND.getStatus().value())
+            .body("errorCode", equalTo(ArticleExceptionType.ARTICLE_NOT_FOUND.getErrorCode()))
+            .body("message", equalTo(ArticleExceptionType.ARTICLE_NOT_FOUND.getErrorMessage()));
     }
 
     @Test
     void 필수값이_비어있을_경우_예외가_발생한다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
+        Map<String, Object> requestBody = Map.of("content", "하나둘셋 얍",
+            "hashtags", List.of("ㄱㄴㄱㄴ", "얍", "모여라"));
+
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
             .pathParam("id", 1)
-            .body(Collections.singletonMap("status", " "))
+            .body(requestBody)
 
             .when()
-            .patch("/community/gathering/{id}")
+            .put("/community/playground/{id}")
 
             .then()
-            .statusCode(INCORRECT_REQUEST_VALUE.getStatus().value())
-            .body("errorCode", equalTo(INCORRECT_REQUEST_VALUE.getErrorCode()))
-            .body("message", equalTo(INCORRECT_REQUEST_VALUE.getErrorMessage()));
-
+            .statusCode(INVALID_REQUEST.getStatus().value())
+            .body("errorCode", equalTo(INVALID_REQUEST.getErrorCode()))
+            .body("message", equalTo(INVALID_REQUEST.getErrorMessage()))
+            .body("detail.title", equalTo("제목은 필수입니다."));
     }
 
     @Test
     void 존재하지_않는_회원인_경우_예외가_발생한다() {
-        String accessToken = jwtProvider.issueAccessToken(99999L);
+        String accessToken = jwtProvider.issueAccessToken(99999999L);
+
+        Map<String, Object> requestBody = Map.of("title", "안녕하세용",
+            "content", "헬륨가스를모곳지", "hashtags",
+            List.of("자유게시판", "개발자좋아효", "코린이"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 9999999)
-            .body(Collections.singletonMap("status", "complete"))
+            .pathParam("id", 10)
+            .body(requestBody)
 
             .when()
-            .patch("/community/gathering/{id}")
+            .put("/community/playground/{id}")
 
             .then()
             .statusCode(MEMBER_NOT_FOUND.getStatus().value())
@@ -166,59 +204,24 @@ class StudyUpdateStatusTest {
     }
 
     @Test
-    void id가_존재하지_않는_경우_예외를_발생시킨다() {
+    void 작성자가_일치하지_않는_경우_예외가_발생한다() {
         String accessToken = jwtProvider.issueAccessToken(1L);
 
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 9999999)
-            .body(Collections.singletonMap("status", "complete"))
-
-            .when()
-            .patch("/community/gathering/{id}")
-
-            .then()
-            .statusCode(ARTICLE_NOT_FOUND.getStatus().value())
-            .body("errorCode", equalTo(ARTICLE_NOT_FOUND.getErrorCode()))
-            .body("message", equalTo(ARTICLE_NOT_FOUND.getErrorMessage()));
-    }
-
-    @Test
-    void 다른_회원의_상태변경을_요청할_경우_예외를_발생시킨다() {
-        String accessToken = jwtProvider.issueAccessToken(2L);
+        Map<String, Object> requestBody = Map.of("title", "안녕하세용",
+            "content", "헬륨가스를모곳지", "hashtags",
+            List.of("자유게시판", "개발자좋아효", "코린이"));
 
         given(documentationSpec)
             .contentType(APPLICATION_JSON_VALUE)
             .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-            .body(Collections.singletonMap("status", "incomplete"))
+            .pathParam("id", 10)
+            .body(requestBody)
 
             .when()
-            .patch("/community/gathering/{id}")
+            .put("/community/playground/{id}")
 
             .then()
             .statusCode(ARTICLE_NOT_MATCHED_WRITER.getStatus().value())
-            .body("errorCode", equalTo(ARTICLE_NOT_MATCHED_WRITER.getErrorCode()))
-            .body("message", equalTo(ARTICLE_NOT_MATCHED_WRITER.getErrorMessage()));
-    }
-
-    @Test
-    void 현재와_같은_상태변경으_요청할_경우_예외가_발생한다() {
-        String accessToken = jwtProvider.issueAccessToken(1L);
-
-        given(documentationSpec)
-            .contentType(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .pathParam("id", 1)
-            .body(Collections.singletonMap("status", "incomplete"))
-
-            .when()
-            .patch("/community/gathering/{id}")
-
-            .then()
-            .statusCode(ALREADY_CHANGED_STATUS.getStatus().value())
-            .body("errorCode", equalTo(ALREADY_CHANGED_STATUS.getErrorCode()))
-            .body("message", equalTo(ALREADY_CHANGED_STATUS.getErrorMessage()));
+            .body("errorCode", equalTo(ARTICLE_NOT_MATCHED_WRITER.getErrorCode()));
     }
 }
