+++ noriter/src/test/java/com/codewarts/noriter/article/docs/question/QuestionDeleteTest.java
@@ -162,7 +162,7 @@ class QuestionDeleteTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("questionRemove.id: ID가 비어있습니다."));
+            .body("message", equalTo("remove.id: ID가 비어있습니다."));
     }
 
     @Test
@@ -180,7 +180,7 @@ class QuestionDeleteTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("questionRemove.id: 게시글 ID는 양수이어야 합니다."));
+            .body("message", equalTo("remove.id: 게시글 ID는 양수이어야 합니다."));
     }
 
     @Test
