+++ noriter/src/test/java/com/codewarts/noriter/article/docs/question/QuestionDetailTest.java
@@ -136,7 +136,7 @@ class QuestionDetailTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("questionDetail.id: 게시글 ID는 양수이어야 합니다."));
+            .body("message", equalTo("getDetail.id: 게시글 ID는 양수이어야 합니다."));
     }
 
     @Test
@@ -152,7 +152,7 @@ class QuestionDetailTest {
             .then()
             .statusCode(CommonExceptionType.INVALID_REQUEST.getStatus().value())
             .body("errorCode", equalTo(CommonExceptionType.INVALID_REQUEST.getErrorCode()))
-            .body("message", equalTo("questionDetail.id: ID가 비어있습니다."));
+            .body("message", equalTo("getDetail.id: ID가 비어있습니다."));
     }
 
     @Test
